refactor: Money와 LottoCount 생성시에 음수 입력 검사
