+++ src/main/java/model/lotto/LottoCount.java
@@ -1,5 +1,7 @@
 package model.lotto;
 
+import utils.InputValidateUtils;
+
 public class LottoCount {
 	public static final String CANT_MAKE_AMOUNT_ERROR_MESSAGE = "[Error]: 이 만큼의 로또를 생성할 수 없습니다.";
 
@@ -7,6 +9,7 @@ public class LottoCount {
 	private int madeCount;
 
 	public LottoCount(int count) {
+		InputValidateUtils.inputNegative(count);
 		this.count = count;
 		madeCount = 0;
 	}
