+++ src/main/java/model/lottonumber/LottoNumbers.java
@@ -2,6 +2,8 @@ package model.lottonumber;
 
 import java.util.List;
 
+import model.lottonumber.generationstrategy.LottoNumbersGenerationStrategy;
+
 public class LottoNumbers {
 	private static final String NUMBER_SIZE_ERROR_MESSAGE = "[Error]: 로또는 %d개의 숫자여야 합니다.";
 	private static final String LOTTO_NUMBER_REDUPLICATION_ERROR_MESSAGE = "[Error]: 로또 번호는 중복이 있으면 안됩니다";
