+++ src/main/java/model/lottonumber/generationstrategy/LottoNumbersGenerationStrategy.java
@@ -1,6 +1,8 @@
-package model.lottonumber;
+package model.lottonumber.generationstrategy;
 
 import java.util.List;
+
+import model.lottonumber.LottoNumber;
 
 public interface LottoNumbersGenerationStrategy {
 	List<LottoNumber> generate(int size);
