+++ src/test/java/model/lotto/LottoTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 
 import model.lottonumber.LottoNumber;
 import model.lottonumber.LottoNumbers;
-import model.lottonumber.LottoNumbersGenerationStrategy;
+import model.lottonumber.generationstrategy.LottoNumbersGenerationStrategy;
 import model.result.Rank;
 import model.winningnumber.WinningLottoNumberDTO;
