+++ src/test/java/model/lottonumber/RandomLottoNumbersGenerationStrategyTest.java
@@ -7,6 +7,8 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import model.lottonumber.generationstrategy.RandomLottoNumbersGenerationStrategy;
+
 public class RandomLottoNumbersGenerationStrategyTest {
 
 	@Test
